chore: codebase audit cleanup
- Remove dead code: xppTools.ts, xppPrompts.ts, debug-parser.ts, patch scripts
- Remove TermRelationshipGraph (write-only) from context, index.ts, suggestionEngine, tests
- Fix 6 drifted tool schemas in mcpServer.ts (search, batch_search, get_class_info, etc.)
- Consolidate 6 scripts/db-tools/*.mjs into single scripts/db-maintenance.mjs CLI
- Add missing vitest.integration.config.ts referenced by test:integration script
- Optimize copilot-instructions.md (849 -> 273 lines)
- Fix run_sys_test alias, missing progress message, prompt contradictions

Build clean, 380/380 tests pass.
