+++ src/prompts/index.ts
@@ -2,6 +2,5 @@
  * Prompts module exports
  */
 
-export * from './xppPrompts.js';
 export * from './codeReview.js';
 export * from './systemInstructions.js';
