+++ src/utils/suggestionEngine.ts
@@ -168,114 +168,6 @@ function generateRelatedSuggestions(
 }
 
 /**
- * Build term relationship graph from symbol metadata
- * Analyzes what terms are commonly used together
- */
-export class TermRelationshipGraph {
-  private relationships: Map<string, Map<string, number>> = new Map();
-  
-  /**
-   * Build relationship graph from symbols
-   */
-  build(symbols: XppSymbol[]): void {
-    for (const symbol of symbols) {
-      const baseTerm = symbol.name.toLowerCase();
-      
-      // Extract related terms from various metadata fields
-      const relatedTerms = this.extractRelatedTerms(symbol);
-      
-      // Record co-occurrences
-      for (const relatedTerm of relatedTerms) {
-        this.addRelationship(baseTerm, relatedTerm);
-      }
-    }
-  }
-  
-  /**
-   * Extract related terms from symbol metadata
-   */
-  private extractRelatedTerms(symbol: XppSymbol): Set<string> {
-    const terms = new Set<string>();
-    
-    // From used_types field
-    if (symbol.usedTypes) {
-      const types = symbol.usedTypes.split(',').map(t => t.trim().toLowerCase());
-      types.forEach(t => terms.add(t));
-    }
-    
-    // From method_calls field
-    if (symbol.methodCalls) {
-      const calls = symbol.methodCalls.split(',').map(c => c.trim().toLowerCase());
-      calls.forEach(c => terms.add(c));
-    }
-    
-    // From related_methods field
-    if (symbol.relatedMethods) {
-      const methods = symbol.relatedMethods.split(',').map(m => m.trim().toLowerCase());
-      methods.forEach(m => terms.add(m));
-    }
-    
-    // From parent class/table
-    if (symbol.parentName) {
-      terms.add(symbol.parentName.toLowerCase());
-    }
-    
-    // From extends class
-    if (symbol.extendsClass) {
-      terms.add(symbol.extendsClass.toLowerCase());
-    }
-    
-    return terms;
-  }
-  
-  /**
-   * Add or increment relationship between two terms
-   */
-  private addRelationship(term1: string, term2: string): void {
-    if (term1 === term2) return;
-    
-    if (!this.relationships.has(term1)) {
-      this.relationships.set(term1, new Map());
-    }
-    
-    const termRelations = this.relationships.get(term1)!;
-    const currentCount = termRelations.get(term2) || 0;
-    termRelations.set(term2, currentCount + 1);
-  }
-  
-  /**
-   * Get related terms for a given term
-   */
-  getRelatedTerms(term: string, limit: number = 5): Array<{ term: string; strength: number }> {
-    const termLower = term.toLowerCase();
-    const relations = this.relationships.get(termLower);
-    
-    if (!relations) return [];
-    
-    // Convert to array and sort by strength
-    return Array.from(relations.entries())
-      .map(([relatedTerm, count]) => ({
-        term: relatedTerm,
-        strength: count
-      }))
-      .sort((a, b) => b.strength - a.strength)
-      .slice(0, limit);
-  }
-  
-  /**
-   * Get total relationship count for a term (popularity indicator)
-   */
-  getTermPopularity(term: string): number {
-    const termLower = term.toLowerCase();
-    const relations = this.relationships.get(termLower);
-    
-    if (!relations) return 0;
-    
-    return Array.from(relations.values()).reduce((sum, count) => sum + count, 0);
-  }
-}
-
-/**
  * Format suggestions for display
  */
 export function formatSuggestions(suggestions: SearchSuggestion[]): string {
