+++ tests/tools/labels.test.ts
@@ -107,7 +107,6 @@ const buildContext = (overrides: Partial<XppServerContext> = {}): XppServerConte
   } as any,
   workspaceScanner: {} as any,
   hybridSearch: {} as any,
-  termRelationshipGraph: {} as any,
   ...overrides,
 });
