+++ tests/tools/method-fallback.test.ts
@@ -67,7 +67,6 @@ const buildContext = (overrides: Partial<XppServerContext> = {}): XppServerConte
   } as any,
   workspaceScanner: {} as any,
   hybridSearch: {} as any,
-  termRelationshipGraph: {} as any,
   ...overrides,
 });
