+++ tests/tools/object-info.test.ts
@@ -165,7 +165,6 @@ const buildContext = (overrides: Partial<XppServerContext> = {}): XppServerConte
   } as any,
   workspaceScanner: {} as any,
   hybridSearch: {} as any,
-  termRelationshipGraph: {} as any,
   ...overrides,
 });
